requirements.txt
Django
Pillow
psycopg2
numpy
pandas
sklearn
matplotlib
beautifulsoup4
djangorestframework==3.9.4 
#djangorestframework
django-filter
django-cors-headers
django-mysql
django-rest-auth
django-allauth
django-vote
